+++ app/routes/frontpage/components/CompactEvents.css
