+++ app/routes/frontpage/components/CompactEvents.tsx
