+++ app/routes/frontpage/components/PublicFrontpage.tsx
@@ -14,9 +14,9 @@ import AuthSection from 'app/components/AuthSection/AuthSection';
 import { Image } from 'app/components/Image';
 import { readmeIfy } from 'app/components/ReadmeLogo';
 import { selectPinned } from 'app/reducers/frontpage';
-import LatestReadme from 'app/routes/overview/components/LatestReadme';
-import Pinned from 'app/routes/overview/components/Pinned';
-import { itemUrl, renderMeta } from 'app/routes/overview/components/utils';
+import LatestReadme from 'app/routes/frontpage/components/LatestReadme';
+import Pinned from 'app/routes/frontpage/components/Pinned';
+import { itemUrl, renderMeta } from 'app/routes/frontpage/components/utils';
 import { useAppDispatch, useAppSelector } from 'app/store/hooks';
 import CompactEvents from './CompactEvents';
 import styles from './PublicFrontpage.css';
