--- app/routes/overview/index.ts
@@ -1,3 +1,0 @@
-import Overview from './components/Overview';
-
-export default Overview;
